requirements.txt
open-clip-torch
streamlit
chromadb==0.5.0
pillow
numpy
tqdm